+++ api-derive/chain/bestNumberLag.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.5.6';
-import type { BlockNumber } from 'https://deno.land/x/polkadot@0.0.9/types/interfaces/index.ts';
+import type { BlockNumber } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
 import type { DeriveApi } from '../types.ts';
 
 import { combineLatest, map } from 'https://esm.sh/rxjs@7.5.6';